Updating to latest mmtk-core and latest Julia

+++ mmtk/src/julia_scanning.rs
@@ -166,7 +166,7 @@ pub unsafe fn scan_julia_object(addr: Address, closure : &mut dyn EdgeVisitor<Ju
         let (mut offset, mut lb, mut ub) = (0, 0, usize::MAX);
         // FIXME: the code below is executed COPY_STACKS has been defined in the C Julia implementation - it is on by default
         if !stkbuf_addr.is_zero() && copy_stack != 0 && (*ta).ptls != std::ptr::null_mut() {
-            if (*ta).tid < 0 {
+            if ((*ta).tid as i16) < 0 {
                 panic!("tid must be positive.")
             }
             let stackbase = ((*UPCALLS).get_stackbase)((*ta).tid);
